feat: mejorar gestión de intercambios y chat
- Se ha añadido la funcionalidad para manejar mensajes en tiempo real en los chats de intercambios mediante la implementación de ApiChatRepository.
- Se ha actualizado la pantalla ChatScreen para utilizar el nuevo repositorio, mejorando la experiencia de usuario al interactuar con los mensajes.
- Se han introducido nuevos métodos en AuthService para obtener el ID del usuario actual, facilitando la identificación de mensajes enviados por el usuario.
- Se ha implementado ExchangeChatReadService para gestionar la visualización de mensajes nuevos en los chats de intercambio.
- Se han realizado ajustes en los modelos y pantallas para soportar la nueva lógica de chat, incluyendo la visualización de nombres de remitentes y la gestión de mensajes no leídos.

Este commit establece una base sólida para la interacción en tiempo real en los chats de intercambios, mejorando la comunicación entre usuarios.

+++ lib/models/user_profile.dart
@@ -17,7 +17,8 @@ class UserProfile {
   final int exchanges;
   final double rating;
   final int languagesCount;
-  final int hoursTotal;
+  /** Horas totales de intercambios completados (puede ser decimal, ej. 1.5). */
+  final double hoursTotal;
   final int currentStreakDays;
   final int bestStreakDays;
   final int medals;
@@ -56,7 +57,7 @@ class UserProfile {
     int? exchanges,
     double? rating,
     int? languagesCount,
-    int? hoursTotal,
+    double? hoursTotal,
     int? currentStreakDays,
     int? bestStreakDays,
     int? medals,
